# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Hotel Management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/feedback/get/getbyid/{id}": {
            "get": {
                "summary": "Get by id feeback info",
                "responses": {
                    "200": {
                        "description": "Feedback found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feedback"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Token or Feedback not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/api/feedback/delete/{id}": {
            "delete": {
                "summary": "Delete feedback by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback deleted"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Token or Feedback not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/api/feedback/patch": {
            "patch": {
                "summary": "Update feedback info",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 100
                                    }
                                },
                                "required": [
                                    "id",
                                    "description"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Feedback"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "403": {
                        "description": "You are not admin"
                    },
                    "404": {
                        "description": "Token or Feedback not found"
                    },
                    "400": {
                        "description": "Feedback description is required or Feedback description maximum 100 characters."
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/feedback/patch/restorebyid/{id}": {
            "patch": {
                "summary": "Restore by id feedback",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback restored"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "403": {
                        "description": "You are not admin"
                    },
                    "404": {
                        "description": "Token or Feedback not found"
                    },
                    "400": {
                        "description": "Feedback already restored"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/feedback/patch/restoreall": {
            "patch": {
                "summary": "Restore all feedbacks",
                "responses": {
                    "200": {
                        "description": "Feedbacks restored"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "403": {
                        "description": "You are not admin"
                    },
                    "404": {
                        "description": "Token or Soft deleted feedbacks not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/feedback/patch/read/{id}": {
            "patch": {
                "summary": "isRead change feedback",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback marked as read or Feedback marked as unread"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "403": {
                        "description": "You are not admin"
                    },
                    "404": {
                        "description": "Token or Feedback not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    }
}