# src/app/api/feedback/patch/read/[id]/route.ts
import { cookies } from "next/headers";
import { NextRequest, NextResponse } from "next/server";
import jwt from "jsonwebtoken";
import mongoose from 'mongoose';
import { MyJwtPayload } from "@/src/types/jwtType";
import Feedback from "@/src/models/feedbackModel";

interface Context {
    params: Promise<{
        id: string;
    }>;
}

const SECRET = process.env.JWT_SECRET!;
export async function PATCH(
    request: NextRequest,
    context: Context
) {
    try {
        const { id } = await context.params;

        const cookieStore = await cookies();
        const token = cookieStore.get("accessToken")?.value;

        if (!token) {
            return NextResponse.json({ message: "Token is not found" }, { status: 404 });
        }

        let decoded;
        try {
            decoded = jwt.verify(token, SECRET) as MyJwtPayload;
        } catch (err) {
            return NextResponse.json({ message: "Token is invalid" }, { status: 401 });
        }

        if (!decoded.isAdmin) {
            return NextResponse.json({ message: "You are not admin" }, { status: 403 });
        }

        const feedbackId = id;
        const readFeedback = await Feedback.findById(feedbackId)
        if (!readFeedback) {
            return NextResponse.json({ message: "Feedback is not found" }, { status: 404 });
        }
        if (readFeedback.isRead) {
            readFeedback.isRead = false;
            await readFeedback.save();
            return NextResponse.json({ message: `Feedback marked as unread` }, { status: 200 });
        } else {
            readFeedback.isRead = true;
            await readFeedback.save();
            return NextResponse.json({ message: `Feedback marked as read` }, { status: 200 });
        }
    } catch (error: unknown) {
        if (error instanceof mongoose.Error.ValidationError) {
            const errors = Object.values(error.errors).map(el => {
                if (el instanceof mongoose.Error.ValidatorError) {
                    return el.message;
                }
                return 'Validation error';
            });
            return NextResponse.json({ error: errors.join(', ') }, { status: 400 });
        } else if (error instanceof Error) {
            return NextResponse.json({ error: error.message }, { status: 500 });
        } else {
            return NextResponse.json({ error: 'Unknown error' }, { status: 500 });
        }
    }
}